stacks/hass-rag/pyproject.toml
[project]
name = "hass-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.6.2",
    "fastapi[standard]>=0.115.6",
    "fastapi-utils>=0.8.0",
    "numpy>=2.2.1",
    "openai>=1.59.7",
    "requests>=2.32.3",
    "typing-inspect>=0.9.0",
    "ollama>=0.4.7",
    "pyyaml>=6.0.2",
]
